Use the standard VS Code approach to enable vendoring
